+++ _smart_splitting.ts
@@ -1,140 +1,20 @@
-import { xts } from "./deps.ts";
+import { bundleImportTransformer } from "./system.ts";
 
-const printer: xts.Printer = xts.createPrinter(
-  { newLine: xts.NewLineKind.LineFeed, removeComments: false },
-);
+import { ts } from "./deps.ts";
 
-function bundleImportTransformer(specifier: string) {
-  return (context: xts.TransformationContext) => {
-    const visit: xts.Visitor = (node: xts.Node) => {
-      if (
-        xts.isCallExpression(node) &&
-        node.expression?.expression?.escapedText === "System" &&
-        node.expression?.name?.escapedText === "register"
-      ) {
-        return xts.visitEachChild(node, (node: xts.Node) => {
-          if (node.kind === xts.SyntaxKind["FunctionExpression"]) {
-            return xts.visitEachChild(node, (node: xts.Node) => {
-              if (node.kind === xts.SyntaxKind["Block"]) {
-                return xts.visitEachChild(node, (node: xts.Node) => {
-                  if (node.kind === xts.SyntaxKind["ReturnStatement"]) {
-                    return xts.visitEachChild(node, (node: xts.Node) => {
-                      if (
-                        node.kind === xts.SyntaxKind["ObjectLiteralExpression"]
-                      ) {
-                        return xts.visitEachChild(node, (node: xts.Node) => {
-                          if (
-                            node.kind === xts.SyntaxKind["PropertyAssignment"]
-                          ) {
-                            return xts.visitEachChild(node, (node: xts.Node) => {
-                              if (
-                                node.kind ===
-                                  xts.SyntaxKind["FunctionExpression"]
-                              ) {
-                                return xts.visitEachChild(
-                                  node,
-                                  (node: xts.Node) => {
-                                    if (node.kind === xts.SyntaxKind["Block"]) {
-                                      return xts.visitEachChild(
-                                        node,
-                                        (node: xts.Node) => {
-                                          if (
-                                            node.kind ===
-                                              xts.SyntaxKind[
-                                                "ExpressionStatement"
-                                              ]
-                                          ) {
-                                            return xts.visitEachChild(
-                                              node,
-                                              (node: xts.Node) => {
-                                                if (
-                                                  node.kind ===
-                                                    xts.SyntaxKind[
-                                                      "CallExpression"
-                                                    ] &&
-                                                  node.expression
-                                                      .escapedText ===
-                                                    "exports_1"
-                                                ) {
-                                                  return xts.visitEachChild(
-                                                    node,
-                                                    (node: xts.Node) => {
-                                                      if (
-                                                        node.kind ===
-                                                          xts.SyntaxKind[
-                                                            "BinaryExpression"
-                                                          ]
-                                                      ) {
-                                                        return xts.createBinary(
-                                                          node.left,
-                                                          xts.createToken(
-                                                            xts.SyntaxKind
-                                                              .EqualsToken,
-                                                          ),
-                                                          xts.createPropertyAccess(
-                                                            xts.createIdentifier(
-                                                              specifier,
-                                                            ),
-                                                            node.right.text,
-                                                          ),
-                                                        );
-                                                      }
-                                                      return node;
-                                                    },
-                                                    context,
-                                                  );
-                                                }
-                                                return node;
-                                              },
-                                              context,
-                                            );
-                                          }
-                                          return node;
-                                        },
-                                        context,
-                                      );
-                                    }
-                                    return node;
-                                  },
-                                  context,
-                                );
-                              }
-                              return node;
-                            }, context);
-                          }
-                          return node;
-                        }, context);
-                      }
-                      return node;
-                    }, context);
-                  }
-                  return node;
-                }, context);
-              }
-              return node;
-            }, context);
-          }
-          return node;
-        }, context);
-      }
-      return xts.visitEachChild(node, visit, context);
-    };
-    return (node: xts.Node) => {
-      return xts.visitNode(node, visit);
-    };
-  };
-}
+const printer = ts.createPrinter({ removeComments: false });
 
 export function injectBundleImport(source: string, specifier: string): string {
-  const sourceFile = xts.createSourceFile(
+  const sourceFile = ts.createSourceFile(
     "x.ts",
     source,
-    xts.ScriptTarget.Latest,
-    true,
+    ts.ScriptTarget.Latest,
   );
-  const result = xts.transform(sourceFile, [bundleImportTransformer(specifier)]);
+  
+  const result = ts.transform(sourceFile, [bundleImportTransformer(specifier)]);
+  
   return printer.printNode(
-    xts.EmitHint.SourceFile,
+    ts.EmitHint.SourceFile,
     result.transformed[0],
     sourceFile,
   );
@@ -144,15 +24,16 @@ export function createModuleImport(
   specifier: string,
   filePath: string,
 ): string {
-  const declaration = xts.createImportDeclaration(
+  const declaration = ts.createImportDeclaration(
     undefined,
     undefined,
-    xts.createImportClause(
+    ts.createImportClause(
       undefined,
-      xts.createNamespaceImport(xts.createIdentifier(specifier)),
+      ts.createNamespaceImport(ts.createIdentifier(specifier)),
       false,
     ),
-    xts.createStringLiteral(filePath),
+    ts.createStringLiteral(filePath),
   );
-  return printer.printNode(xts.EmitHint.Unspecified, declaration, undefined);
+  const sourceFile = ts.createSourceFile("x.ts", "", ts.ScriptTarget.Latest);
+  return printer.printNode(ts.EmitHint.Unspecified, declaration, sourceFile);
 }
