+++ plugins/transformers/typescript_inject_specifiers.ts
@@ -1,7 +1,6 @@
-import { ImportMap, path, xts } from "../../deps.ts";
+import { ImportMap, path, ts, xts } from "../../deps.ts";
 import { FileMap, getOutput, Graph } from "../../graph.ts";
 import {
-  CompilerOptions,
   getDynamicImportNode,
   getExportNode,
   getImportNode,
@@ -90,7 +89,7 @@ function injectOutputsTranformer(
 
 interface Config {
   test?: PluginTest;
-  compilerOptions?: CompilerOptions;
+  compilerOptions?: ts.CompilerOptions;
 }
 
 export function typescriptInjectSpecifiers(
